ULL Data Type Changes
  * Updated four occurrences of old frequency datatype to
    unsigned long long; this fixes the mishandling of freqs
    greater than 2 GHz in the serial loop.

+++ src/UI/KX3_ui.cxx
@@ -47,7 +47,7 @@ extern std::queue<XCVR_STATE> queB;
 
 void read_KX3_vfo()
 {
-	unsigned long int freq;
+	unsigned long long freq;
 	freq = selrig->get_vfoA();
 	if (freq != vfoA.freq) {
 		vfoA.freq = freq;
